Move react source around
