--- static/js/views/hello_world.jsx
@@ -1,15 +1,0 @@
-var React = require('react')
-var Device = require('./device_interface.jsx')
-
-var HelloWorld = React.createClass({
-    render: function(){
-        return (
-        <div>
-            <h1>Hello, {this.props.name}!</h1>
-            <Device state='on'/>
-        </div>
-        );
-    }
-});
-
-module.exports = HelloWorld;
